+++ indico/modules/users/blueprint.py
@@ -11,10 +11,11 @@ from flask import request
 
 from indico.modules.users.api import RHUserFavoritesAPI, fetch_authenticated_user
 from indico.modules.users.controllers import (RHAcceptRegistrationRequest, RHAdmins, RHExportDashboardICS,
-                                              RHPersonalData, RHRegistrationRequestList, RHRejectRegistrationRequest,
-                                              RHUserBlock, RHUserDashboard, RHUserEmails, RHUserEmailsDelete,
-                                              RHUserEmailsSetPrimary, RHUserEmailsVerify, RHUserFavorites,
-                                              RHUserFavoritesCategoryAPI, RHUserFavoritesUserRemove,
+                                              RHPersonalData, RHProfilePicture, RHProfilePictureDelete,
+                                              RHProfilePictureDisplay, RHRegistrationRequestList,
+                                              RHRejectRegistrationRequest, RHUserBlock, RHUserDashboard, RHUserEmails,
+                                              RHUserEmailsDelete, RHUserEmailsSetPrimary, RHUserEmailsVerify,
+                                              RHUserFavorites, RHUserFavoritesCategoryAPI, RHUserFavoritesUserRemove,
                                               RHUserFavoritesUsersAdd, RHUserPreferences, RHUsersAdmin,
                                               RHUsersAdminCreate, RHUsersAdminMerge, RHUsersAdminMergeCheck,
                                               RHUsersAdminSettings, RHUserSearch, RHUserSearchInfo,
@@ -46,6 +47,9 @@ with _bp.add_prefixed_rules('/<int:user_id>'):
     _bp.add_url_rule('/suggestions/categories/<int:category_id>', 'user_suggestions_remove', RHUserSuggestionsRemove,
                      methods=('DELETE',))
     _bp.add_url_rule('/profile/', 'user_profile', RHPersonalData, methods=('GET', 'POST'))
+    _bp.add_url_rule('/profile/picture-<slug>.png', 'profile_picture_display', RHProfilePictureDisplay)
+    _bp.add_url_rule('/profile/picture', 'upload_profile_picture', RHProfilePicture, methods=('POST',))
+    _bp.add_url_rule('/profile/picture', 'delete_profile_picture', RHProfilePictureDelete, methods=('DELETE',))
     _bp.add_url_rule('/preferences/', 'user_preferences', RHUserPreferences, methods=('GET', 'POST'))
     _bp.add_url_rule('/favorites/', 'user_favorites', RHUserFavorites)
     _bp.add_url_rule('/favorites/users/', 'user_favorites_users_add', RHUserFavoritesUsersAdd, methods=('POST',))