+++ indico/modules/users/controllers.py
@@ -7,6 +7,7 @@
 
 from __future__ import unicode_literals
 
+import os
 from collections import namedtuple
 from io import BytesIO
 from operator import attrgetter, itemgetter
@@ -15,6 +16,7 @@ from dateutil.relativedelta import relativedelta
 from flask import flash, jsonify, redirect, request, session
 from markupsafe import Markup, escape
 from marshmallow import fields
+from PIL import Image
 from sqlalchemy.orm import joinedload, load_only, subqueryload, undefer
 from sqlalchemy.orm.exc import StaleDataError
 from webargs import validate
@@ -24,6 +26,7 @@ from indico.core import signals
 from indico.core.auth import multipass
 from indico.core.db import db
 from indico.core.db.sqlalchemy.util.queries import get_n_matching
+from indico.core.errors import UserValueError
 from indico.core.marshmallow import mm
 from indico.core.notifications import make_email, send_email
 from indico.legacy.common.cache import GenericCache
@@ -36,18 +39,21 @@ from indico.modules.events import Event
 from indico.modules.events.util import serialize_event_for_ical
 from indico.modules.users import User, logger, user_management_settings
 from indico.modules.users.forms import (AdminAccountRegistrationForm, AdminsForm, AdminUserSettingsForm, MergeForm,
-                                        SearchForm, UserDetailsForm, UserEmailsForm, UserPreferencesForm)
+                                        SearchForm, UserDetailsForm, UserEmailsForm, UserPictureForm,
+                                        UserPreferencesForm)
 from indico.modules.users.models.emails import UserEmail
 from indico.modules.users.operations import create_user
-from indico.modules.users.util import (get_linked_events, get_related_categories, get_suggested_categories, merge_users,
-                                       search_users, serialize_user)
+from indico.modules.users.util import (get_linked_events, get_picture_data, get_related_categories,
+                                       get_suggested_categories, merge_users, search_users, serialize_user)
 from indico.modules.users.views import WPUser, WPUserDashboard, WPUsersAdmin
 from indico.util.date_time import now_utc
 from indico.util.event import truncate_path
+from indico.util.fs import secure_filename
 from indico.util.i18n import _
+from indico.util.images import square
 from indico.util.marshmallow import HumanizedDate, validate_with_message
 from indico.util.signals import values_from_signal
-from indico.util.string import make_unique_token
+from indico.util.string import crc32, make_unique_token
 from indico.web.args import use_kwargs
 from indico.web.flask.templating import get_template_module
 from indico.web.flask.util import send_file, url_for
@@ -177,13 +183,65 @@ class RHPersonalData(RHUserBase):
     def _process(self):
         form = UserDetailsForm(obj=FormDefaults(self.user, skip_attrs={'title'}, title=self.user._title),
                                synced_fields=self.user.synced_fields, synced_values=self.user.synced_values)
+        pic_form = UserPictureForm()
         if form.validate_on_submit():
             self.user.synced_fields = form.synced_fields
             form.populate_obj(self.user, skip=self.user.synced_fields)
             self.user.synchronize_data(refresh=True)
             flash(_('Your personal data was successfully updated.'), 'success')
             return redirect(url_for('.user_profile'))
-        return WPUser.render_template('personal_data.html', 'personal_data', user=self.user, form=form)
+        elif self.user.has_picture:
+            pic_form.picture.data = self.user
+        return WPUser.render_template('personal_data.html', 'personal_data', user=self.user, form=form,
+                                      pic_form=pic_form)
+
+
+class RHProfilePictureDisplay(RHUserBase):
+    def _process(self):
+        if not self.user.has_picture:
+            raise NotFound
+        metadata = self.user.picture_metadata
+        return send_file('user.png', BytesIO(self.user.picture), mimetype=metadata['content_type'],
+                         conditional=True)
+
+
+class RHProfilePicture(RHUserBase):
+    def _process(self):
+        f = request.files['picture']
+        try:
+            pic = Image.open(f)
+        except IOError:
+            raise UserValueError(_('You cannot upload this file as profile picture.'))
+        if pic.format.lower() not in {'jpeg', 'png', 'gif'}:
+            raise UserValueError(_('The file has an invalid format ({format}).').format(format=pic.format))
+        if pic.mode not in {'RGB', 'RGBA'}:
+            pic = pic.convert('RGB')
+        image_bytes = BytesIO()
+        pic = square(pic)
+        if pic.height > 256:
+            pic = pic.resize((256, 256), resample=Image.BICUBIC)
+        pic.save(image_bytes, 'PNG')
+        image_bytes.seek(0)
+        content = image_bytes.read()
+        self.user.picture = content
+        self.user.picture_metadata = {
+            'hash': crc32(content),
+            'size': len(content),
+            'filename': os.path.splitext(secure_filename(f.filename, 'user'))[0] + '.png',
+            'content_type': 'image/png'
+        }
+        flash(_('Profile picture uploaded'), 'success')
+        logger.info('Profile picture of user %s uploaded by %s', self.user, session.user)
+        return jsonify_data(content=get_picture_data(self.user))
+
+
+class RHProfilePictureDelete(RHUserBase):
+    def _process(self):
+        self.user.picture = None
+        self.user.picture_metadata = None
+        flash(_('Profile picture deleted'), 'success')
+        logger.info('Profile picture of user %s deleted by %s', self.user, session.user)
+        return jsonify_data(content=None)
 
 
 class RHUserPreferences(RHUserBase):